pyproject.toml
[tool.poetry]
name = "keycloak-manager"
version = "0.1.0"
description = ""
authors = ["Abdullah Alqahtani <0xqahta0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
